requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0.0  # PyTorch requires NumPy 1.x

# Data handling and preprocessing
pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization and evaluation
matplotlib>=3.7.0
seaborn>=0.13.0
opencv-python>=4.8.0

# Grad-CAM for model interpretability
grad-cam>=1.5.0

# Web application
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Configuration
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Progress bars
tqdm>=4.66.0

# ML experiment tracking
mlflow>=2.10.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For async test client
ruff>=0.8.0
mypy>=1.8.0
pre-commit>=3.6.0

# Installation notes:
# For GPU (CUDA) support, install PyTorch separately first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Then install the rest:
#   pip install -r requirements.txt
#
# For CPU-only:
#   pip install -r requirements.txt
